refactor(suggest): move Suggestions response type to suggest package
