+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/request/suggest/Suggestions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2020-2022 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.domain;
+package com.b2international.snowowl.core.request.suggest;
 
 import java.util.Collections;
 import java.util.List;
+
+import com.b2international.snowowl.core.domain.Concept;
+import com.b2international.snowowl.core.domain.PageableCollectionResource;
 
 /**
  * @since 7.7
